LayoutTests/fast/dom/CSSStyleDeclaration/access-longest-css-property-expected.txt

The test parsing of the boundary value by checking the longest CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.body).webkitMatchNearestMailBlockquoteColor is "match"
PASS document.getElementById("target").style.webkitMatchNearestMailBlockquoteColor is "match"
PASS document.body.style.webkitMatchNearestMailBlockquoteColor is ""
PASS successfullyParsed is true

TEST COMPLETE
